refactor: extract task-board operations module (issue #385)
